requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sse-starlette
duckduckgo-search
requests
playwright
beautifulsoup4
readability-lxml
google-generativeai
sentence-transformers
chromadb
tiktoken
lxml
